+++ 0x0C-more_malloc_free/main.h
@@ -2,6 +2,6 @@
 #define MAIN_H
 
 int _putchar(char c);
-void *malloc_checked(unsigned int b);
+void *malloc_checked(unsigned long int b);
 
 #endif
